Add Experiment inheritance and parent/child relationships
Experiments can now be initialized using an existing Experiment object and will
inherit the DGPs, Methods, and Evaluators of the parent. The basic vignette is
updated to show the new functionality. This commit also refactors a couple of
things in the Experiment class:

* {dgp,method,evaluator}_list are now private and accessed via public getter
methods (e.g., get_dgps), which recursively access ancestors as well.
* {dgp,method,evaluator} add and update methods now check that inputs have the
right class, as does Experiment initialization.