Catch and return errors from all parallel workers in Experiment$fit
Also return partial results from DGP+Method combos that were successful before
the worker encountered the first error. The error object signaled by
Experiment$fit contains a tibble of the partial results and errors, and each
error in the tibble contains the input data and the name of the DGP/Method that
was being processed when the error occurred.